mcp server functionality merged with existing workflow using http transport
