+++ pkg/cmd/costmodel/costmodel.go
@@ -14,12 +14,14 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/rs/cors"
 
+	mcp_sdk "github.com/modelcontextprotocol/go-sdk/mcp"
 	"github.com/opencost/opencost/core/pkg/errors"
 	"github.com/opencost/opencost/core/pkg/log"
 	"github.com/opencost/opencost/core/pkg/version"
 	"github.com/opencost/opencost/pkg/costmodel"
 	"github.com/opencost/opencost/pkg/env"
 	"github.com/opencost/opencost/pkg/filemanager"
+	opencost_mcp "github.com/opencost/opencost/pkg/mcp"
 	"github.com/opencost/opencost/pkg/metrics"
 )
 
@@ -69,6 +71,16 @@ func Execute(conf *Config) error {
 	// valid for CustomCostPipelineService to be nil
 	router.GET("/customCost/status", customCostPipelineService.GetCustomCostStatusHandler())
 
+	// Initialize MCP Server if enabled and Kubernetes is available
+	if conf.MCPServerEnabled && a != nil {
+		err := StartMCPServer(context.Background(), a)
+		if err != nil {
+			log.Errorf("Failed to start MCP server: %v", err)
+		}
+	} else if conf.MCPServerEnabled {
+		log.Warnf("MCP Server is enabled but Kubernetes is not available. MCP server requires Kubernetes to function.")
+	}
+
 	apiutil.ApplyContainerDiagnosticEndpoints(router)
 
 	rootMux := http.NewServeMux()
@@ -99,7 +111,7 @@ func StartExportWorker(ctx context.Context, model costmodel.AllocationModel) err
 			select {
 			case <-ctx.Done():
 				return
-			case <-time.After(nextRunAt.Sub(time.Now())):
+			case <-time.After(time.Until(nextRunAt)):
 				err := costmodel.UpdateCSV(ctx, fm, model, env.GetExportCSVLabelsAll(), env.GetExportCSVLabelsList())
 				if err != nil {
 					// it's background worker, log error and carry on, maybe next time it will work
@@ -113,4 +125,192 @@ func StartExportWorker(ctx context.Context, model costmodel.AllocationModel) err
 		}
 	}()
 	return nil
+}
+
+// StartMCPServer starts the MCP server as a background service
+func StartMCPServer(ctx context.Context, accesses *costmodel.Accesses) error {
+	if accesses == nil {
+		return fmt.Errorf("cost model accesses not available")
+	}
+
+	log.Info("Initializing MCP server...")
+
+	// Create MCP server using existing OpenCost dependencies
+	mcpServer := opencost_mcp.NewMCPServer(accesses.Model, accesses.CloudProvider, nil)
+
+	// Create MCP SDK server
+	sdkServer := mcp_sdk.NewServer(&mcp_sdk.Implementation{
+		Name:    "opencost-mcp-server",
+		Version: "v1.0.0",
+	}, nil)
+
+	// Define tool handlers
+	handleAllocationCosts := func(ctx context.Context, req *mcp_sdk.CallToolRequest, args AllocationArgs) (*mcp_sdk.CallToolResult, interface{}, error) {
+		// Parse step duration if provided
+		var step time.Duration
+		var err error
+		if args.Step != "" {
+			step, err = time.ParseDuration(args.Step)
+			if err != nil {
+				return nil, nil, fmt.Errorf("invalid step duration '%s': %w", args.Step, err)
+			}
+		}
+
+		queryRequest := &opencost_mcp.OpenCostQueryRequest{
+			QueryType: opencost_mcp.AllocationQueryType,
+			Window:    args.Window,
+			AllocationParams: &opencost_mcp.AllocationQuery{
+				Step:                                  step,
+				Accumulate:                            args.Accumulate,
+				ShareIdle:                             args.ShareIdle,
+				Aggregate:                             args.Aggregate,
+				IncludeIdle:                           args.IncludeIdle,
+				IdleByNode:                            args.IdleByNode,
+				IncludeProportionalAssetResourceCosts: args.IncludeProportionalAssetResourceCosts,
+				IncludeAggregatedMetadata:             args.IncludeAggregatedMetadata,
+				ShareLB:                               args.ShareLB,
+			},
+		}
+
+		mcpReq := &opencost_mcp.MCPRequest{
+			Query: queryRequest,
+		}
+
+		mcpResp, err := mcpServer.ProcessMCPRequest(mcpReq)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to process allocation request: %w", err)
+		}
+
+		return nil, mcpResp, nil
+	}
+
+	handleAssetCosts := func(ctx context.Context, req *mcp_sdk.CallToolRequest, args AssetArgs) (*mcp_sdk.CallToolResult, interface{}, error) {
+		queryRequest := &opencost_mcp.OpenCostQueryRequest{
+			QueryType:   opencost_mcp.AssetQueryType,
+			Window:      args.Window,
+			AssetParams: &opencost_mcp.AssetQuery{},
+		}
+
+		mcpReq := &opencost_mcp.MCPRequest{
+			Query: queryRequest,
+		}
+
+		mcpResp, err := mcpServer.ProcessMCPRequest(mcpReq)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to process asset request: %w", err)
+		}
+
+		return nil, mcpResp, nil
+	}
+
+	handleCloudCosts := func(ctx context.Context, req *mcp_sdk.CallToolRequest, args CloudCostArgs) (*mcp_sdk.CallToolResult, interface{}, error) {
+		queryRequest := &opencost_mcp.OpenCostQueryRequest{
+			QueryType: opencost_mcp.CloudCostQueryType,
+			Window:    args.Window,
+			CloudCostParams: &opencost_mcp.CloudCostQuery{
+				Aggregate:  args.Aggregate,
+				Accumulate: args.Accumulate,
+				Filter:     args.Filter,
+				Provider:   args.Provider,
+				Service:    args.Service,
+				Category:   args.Category,
+				Region:     args.Region,
+				Account:    args.Account,
+			},
+		}
+
+		mcpReq := &opencost_mcp.MCPRequest{
+			Query: queryRequest,
+		}
+
+		mcpResp, err := mcpServer.ProcessMCPRequest(mcpReq)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to process cloud cost request: %w", err)
+		}
+
+		return nil, mcpResp, nil
+	}
+
+	// Register tools
+	mcp_sdk.AddTool(sdkServer, &mcp_sdk.Tool{
+		Name:        "get_allocation_costs",
+		Description: "Retrieves allocation cost data.",
+	}, handleAllocationCosts)
+
+	mcp_sdk.AddTool(sdkServer, &mcp_sdk.Tool{
+		Name:        "get_asset_costs",
+		Description: "Retrieves asset cost data.",
+	}, handleAssetCosts)
+
+	mcp_sdk.AddTool(sdkServer, &mcp_sdk.Tool{
+		Name:        "get_cloud_costs",
+		Description: "Retrieves cloud cost data.",
+	}, handleCloudCosts)
+
+	// Create HTTP handler
+	handler := mcp_sdk.NewStreamableHTTPHandler(func(r *http.Request) *mcp_sdk.Server {
+		return sdkServer
+	}, &mcp_sdk.StreamableHTTPOptions{
+		JSONResponse: true,
+	})
+
+	// Add logging middleware
+	loggingHandler := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		log.Infof("MCP HTTP request: %s %s from %s", req.Method, req.URL.Path, req.RemoteAddr)
+		handler.ServeHTTP(w, req)
+	})
+
+	// Start HTTP server on configured port
+	port := env.GetMCPHTTPPort()
+	log.Infof("Starting MCP HTTP server on port %d...", port)
+
+	server := &http.Server{
+		Addr:    fmt.Sprintf(":%d", port),
+		Handler: loggingHandler,
+	}
+
+	// Start server in a goroutine
+	go func() {
+		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
+			log.Errorf("MCP server failed: %v", err)
+		}
+	}()
+
+	log.Info("MCP server started successfully")
+	return nil
+}
+
+// Tool argument structures for MCP server
+type AllocationArgs struct {
+	Window    string `json:"window"`
+	Aggregate string `json:"aggregate"`
+
+	// Allocation query parameters
+	Step                                  string `json:"step,omitempty"`
+	Resolution                            string `json:"resolution,omitempty"`
+	Accumulate                            bool   `json:"accumulate,omitempty"`
+	ShareIdle                             bool   `json:"share_idle,omitempty"`
+	IncludeIdle                           bool   `json:"include_idle,omitempty"`
+	IdleByNode                            bool   `json:"idle_by_node,omitempty"`
+	IncludeProportionalAssetResourceCosts bool   `json:"include_proportional_asset_resource_costs,omitempty"`
+	IncludeAggregatedMetadata             bool   `json:"include_aggregated_metadata,omitempty"`
+	ShareLB                               bool   `json:"share_lb,omitempty"`
+}
+
+type AssetArgs struct {
+	Window string `json:"window"`
+}
+
+type CloudCostArgs struct {
+	Window    string `json:"window"`
+	Aggregate string `json:"aggregate"`
+
+	// Cloud cost query parameters
+	Accumulate string `json:"accumulate,omitempty"`
+	Filter     string `json:"filter,omitempty"`
+	Provider   string `json:"provider,omitempty"`
+	Service    string `json:"service,omitempty"`
+	Category   string `json:"category,omitempty"`
+	Region     string `json:"region,omitempty"`
+	Account    string `json:"account,omitempty"`
 }
